Function: createGrid
FIRST PART:

To create the grid i iterated 16 times to create all the divs.

The hardest part was how I was going to add the colors and remove them. I knew that I had to use "add" and "remove" but then I remembered one time using "contains" which made it work.

Lesson:
- I needed to use 2 mouseover. One for making the divs change color, and one for "getting" them back to original color on the way back.